requirements.txt
absl-py==2.1.0
altair==5.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
caer==2.0.8
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.4
colorama==0.4.6
comtypes==1.4.7
contourpy==1.3.0
cycler==0.12.1
dlib @ file:///F:/ML%20Projts/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.0
MouseInfo==0.1.3
mpmath==1.3.0
mypy==1.10.1
mypy-extensions==1.0.0
narwhals==1.8.4
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
protobuf==4.25.5
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
PyAutoGUI==0.9.54
pycaw==20240210
pycparser==2.22
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.1.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
pywin32==306
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
setuptools==75.1.0
setuptools-rust==1.10.2
six==1.16.0
smmap==5.0.1
sounddevice==0.5.0
streamlit==1.38.0
streamlit-autorefresh==1.0.1
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.4.1+cpu
torchaudio==2.4.1+cpu
torchvision==0.19.1+cpu
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.2.103
ultralytics-thop==2.0.8
urllib3==2.2.3
watchdog==4.0.2
Werkzeug==3.0.3
wheel==0.44.0
